"인수 테스트 기반 TDD" 2단계 - 경로 조회 기능 (#70)
* feat: 응답에 대한 Custom exception 로직 적용

* refactor: Line과 관련된 Sections, Section, Distance 엔티티 개선

* test: Line과 관련된 엔티티 개선 후, 관련된 엔티티 테스트 코드 추가

* refactor: Distance의 적절한 원시 자료형 변경

* docs(README.md): 스탭별 요구사항 추가

* feat(PathAcceptanceTest.java): 최단 경로 조회 인수테스트 추가

* feat: `최단 경로 조회` 구현에 필요한 코드 포맷 셋팅

* feat: `최단 경로 조회` 기능에 필요한 Path Service 구현

* feat(PathFinder): 경로 찾기에 관련된 기능 `PathFinder` 구현

* feat: "출발역과 도착역이 같은 경우" 예외 상황 처리

* feat: "출발역과 도착역이 연결이 되어 있지 않은 경우" 예외 상황 처리

* feat: "존재하지 않은 출발역이나 도착역을 조회 할 경우" 예외 상황 처리

* refactor: 중복된 Station 조회 메서드 제거

* refactor: 지하철역을 가져오는 불필요한 이중 쿼리 제거

* add(GlobalExceptionHandler.java): 디폴트 전역 예외 처리 핸들러 추가

+++ src/main/java/nextstep/subway/line/dto/LineRequest.java
@@ -7,12 +7,12 @@ public class LineRequest {
     private String color;
     private Long upStationId;
     private Long downStationId;
-    private int distance;
+    private long distance;
 
     public LineRequest() {
     }
 
-    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
+    public LineRequest(String name, String color, Long upStationId, Long downStationId, long distance) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
@@ -36,7 +36,7 @@ public class LineRequest {
         return downStationId;
     }
 
-    public int getDistance() {
+    public long getDistance() {
         return distance;
     }
